Fixed Profile Photos Implementation
Added a new drive for saving profile pictures in the public directory. Azure does not support symlinks so was not able to access the photos from the storage folder directly. Instead had to create a new drive in the config/filesystems.php config file. This drive is in the public directory and photos are stored in a "profile_pics" directory.